requirements.txt
maxapi[webhook]>=0.9.7
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
asyncio
logging
httpx>=0.27.0,<1.0
pydantic>=2.4.0,<3